backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
Flask==3.1.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
